pyproject.toml
[project]
name = "msm-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse (>=1.4.0,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "openai (>=1.66.5,<2.0.0)",
    "sounddevice (>=0.5.1,<0.6.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "pynput (>=1.8.1,<2.0.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "google-genai (>=1.7.0,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "chromadb (>=0.6.3,<0.7.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.1.2"
isort = "^6.0.1"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = ["__pycache__", ".venv", "build", "dist"]

